feat: CLI polish, discovery split, README, tests, clippy (round 3)
A. Error handling & resilience:
- Verbose list output with aligned columns (replaces raw Debug)
- QueueHeader reads via ShmemConf open+drop (no FD leak)

B. TUI polish:
- Segment/app count in title bar
- Filter-aware empty state message
- Queue fill bar in detail view (ASCII: ████░░░░)

C. CLI polish:
- Aligned column output in list command
- --app filter on list, stats commands
- Formatted inspect matching TUI detail view layout
- Colored output (green/red/yellow) when stdout is terminal

D. Performance:
- Flink mtime fast-path in populate_from_fs
- Metadata backfill for previously-stubbed entries

E. Documentation & organization:
- README.md for flux-ctl (210 lines: features, CLI examples,
  TUI keybindings, architecture)
- Split discovery.rs (685 lines) into submodules:
  discovery/{mod,registry,inspect,cli}.rs with pub use re-exports

F. Testing (6 new tests, 114 total):
- empty_state, filter, sort_mode, home_end, page_up_down,
  stress_test (50 groups × 5 segments), compact idempotent,
  entry_by_index bounds, backing_file_size

All clippy warnings fixed (map_or→is_none_or, collapsible_if,
unused import, literal format strings).
