--- crates/flux-ctl/src/discovery.rs
@@ -1,685 +1,0 @@
-//! Discovery and inspection of shared memory segments via the global registry.
-//!
-//! Provides the backing logic for all `flux-ctl` CLI commands (`list`, `inspect`,
-//! `clean`, `scan`) and the TUI data model. Opens the [`ShmemRegistry`], enriches
-//! entries with queue stats, poison detection, and per-PID process info, and
-//! exposes helpers for cleanup and liveness checks.
-
-use std::io::IsTerminal;
-use std::path::Path;
-use std::sync::atomic::{AtomicU64, Ordering};
-
-use crossterm::style::Stylize;
-use serde::Serialize;
-
-use flux_communication::array::ArrayHeader;
-use flux_communication::queue::QueueHeader;
-use flux_communication::registry::{
-    REGISTRY_FLINK_NAME, ShmemEntry, ShmemKind, ShmemRegistry, cleanup_flink,
-};
-use shared_memory::ShmemConf;
-
-#[derive(Clone, Debug)]
-pub struct PidInfo {
-    pub pid: u32,
-    pub alive: bool,
-    pub name: String,
-    pub cmdline: String,
-    pub start_time: String,
-}
-
-pub fn pid_info(pid: u32) -> PidInfo {
-    let alive = is_pid_alive(pid);
-    if !alive {
-        return PidInfo {
-            pid,
-            alive: false,
-            name: String::new(),
-            cmdline: String::new(),
-            start_time: String::new(),
-        };
-    }
-
-    let name = std::fs::read_to_string(format!("/proc/{pid}/comm"))
-        .map(|s| s.trim().to_string())
-        .unwrap_or_default();
-
-    let cmdline = std::fs::read(format!("/proc/{pid}/cmdline"))
-        .map(|bytes| {
-            bytes
-                .split(|&b| b == 0)
-                .filter(|s| !s.is_empty())
-                .map(|s| String::from_utf8_lossy(s).into_owned())
-                .collect::<Vec<_>>()
-                .join(" ")
-        })
-        .unwrap_or_default();
-
-    let start_time = read_start_time(pid).unwrap_or_default();
-
-    PidInfo { pid, alive, name, cmdline, start_time }
-}
-
-pub fn pids_info(entry: &ShmemEntry) -> Vec<PidInfo> {
-    entry.pids.active_pids().iter().map(|&pid| pid_info(pid)).collect()
-}
-
-/// Open the global registry, sweeping dead PIDs and populating from the
-/// filesystem (so pre-existing shmem created before the registry is visible).
-/// Stale flinks whose backing shmem no longer exists are removed.
-pub fn open_registry(base_dir: &Path) -> Option<&'static ShmemRegistry> {
-    let registry_path = base_dir.join(REGISTRY_FLINK_NAME);
-    let reg = ShmemRegistry::open(&registry_path)?;
-    reg.sweep_dead_pids();
-    reg.populate_from_fs(base_dir);
-    Some(reg)
-}
-
-/// Create the registry if needed, then scan the filesystem for pre-existing
-/// shmem and register any missing entries. Stale flinks are removed.
-pub fn scan(base_dir: &Path) -> Result<(), Box<dyn std::error::Error>> {
-    let reg = ShmemRegistry::open_or_create(base_dir);
-    reg.sweep_dead_pids();
-    let result = reg.populate_from_fs(base_dir);
-    println!("Scanned {}", base_dir.display());
-    println!("  Registered:     {}", result.registered);
-    println!("  Already known:  {}", result.already_known);
-    println!("  Stale removed:  {}", result.stale_removed);
-    Ok(())
-}
-
-pub use flux_communication::registry::is_pid_alive;
-
-pub fn app_names(registry: &ShmemRegistry) -> Vec<String> {
-    let mut names: Vec<String> = registry
-        .entries()
-        .iter()
-        .filter(|e| !e.is_empty())
-        .map(|e| e.app_name.as_str().to_string())
-        .collect();
-    names.sort();
-    names.dedup();
-    names
-}
-
-pub fn list_all(base_dir: &Path, verbose: bool) -> Result<(), Box<dyn std::error::Error>> {
-    let Some(registry) = open_registry(base_dir) else {
-        println!(
-            "No flux shmem registry found at {}",
-            base_dir.join(REGISTRY_FLINK_NAME).display()
-        );
-        return Ok(());
-    };
-
-    let entries: Vec<&ShmemEntry> = registry.entries().iter().filter(|e| entry_visible(e)).collect();
-    if entries.is_empty() {
-        println!("No active segments found");
-        return Ok(());
-    }
-
-    let color = std::io::stdout().is_terminal();
-    let apps = app_names(registry);
-    println!("Found {} segments across {} apps\n", entries.len(), apps.len());
-
-    for app in &apps {
-        let app_entries: Vec<&&ShmemEntry> =
-            entries.iter().filter(|e| e.app_name.as_str() == app).collect();
-        if app_entries.is_empty() {
-            continue;
-        }
-        println!("📦 {} ({} segments)", app, app_entries.len());
-        for entry in &app_entries {
-            let alive = entry.pids.any_alive();
-            let poison = check_poison(entry);
-            let status = if poison.is_some() {
-                "☠"
-            } else if alive {
-                "🟢"
-            } else {
-                "💀"
-            };
-            let pids = format_pids(entry);
-            let status_text = format!(
-                "  {} {:14} {:>24}  elem={}B  cap={}  {}",
-                status,
-                entry.kind,
-                entry.type_name.as_str(),
-                entry.elem_size,
-                entry.capacity,
-                pids,
-            );
-            if color {
-                if poison.is_some() {
-                    println!("{}", status_text.yellow());
-                } else if alive {
-                    println!("{}", status_text.green());
-                } else {
-                    println!("{}", status_text.red());
-                }
-            } else {
-                println!("{status_text}");
-            }
-            if let Some(ref pi) = poison {
-                println!(
-                    "    ☠ POISONED: {}/{} slots (first at slot {})",
-                    pi.n_poisoned, pi.total_slots, pi.first_slot,
-                );
-            }
-            if verbose {
-                println!("    flink: {}", entry.flink.as_str());
-                if entry.kind == ShmemKind::Queue {
-                    match QueueHeader::open_shared(entry.flink.as_str()) {
-                        Ok(header) => {
-                            println!(
-                                "    queue: type={:?}  writes={}  capacity={}",
-                                header.queue_type,
-                                header.count.load(Ordering::Relaxed),
-                                header.mask + 1,
-                            );
-                        }
-                        Err(e) => println!("    queue: (could not open: {e})"),
-                    }
-                }
-            }
-        }
-        println!();
-    }
-    Ok(())
-}
-
-// ─── JSON output ────────────────────────────────────────────────────────────
-
-#[derive(Serialize)]
-struct SegmentJson {
-    app: String,
-    kind: String,
-    type_name: String,
-    elem_size: usize,
-    capacity: usize,
-    flink: String,
-    pids: Vec<u32>,
-    alive: bool,
-    poisoned: bool,
-}
-
-impl SegmentJson {
-    fn from_entry(entry: &ShmemEntry) -> Self {
-        let alive = entry.pids.any_alive();
-        let poisoned = check_poison(entry).is_some();
-        Self {
-            app: entry.app_name.as_str().to_string(),
-            kind: format!("{}", entry.kind),
-            type_name: entry.type_name.as_str().to_string(),
-            elem_size: entry.elem_size,
-            capacity: entry.capacity,
-            flink: entry.flink.as_str().to_string(),
-            pids: entry.pids.active_pids(),
-            alive,
-            poisoned,
-        }
-    }
-}
-
-pub fn list_json(base_dir: &Path) -> Result<(), Box<dyn std::error::Error>> {
-    let Some(registry) = open_registry(base_dir) else {
-        println!("[]");
-        return Ok(());
-    };
-
-    let segments: Vec<SegmentJson> = registry
-        .entries()
-        .iter()
-        .filter(|e| entry_visible(e))
-        .map(SegmentJson::from_entry)
-        .collect();
-
-    println!("{}", serde_json::to_string_pretty(&segments)?);
-    Ok(())
-}
-
-// ─── Stats ──────────────────────────────────────────────────────────────────
-
-pub fn stats(base_dir: &Path) -> Result<(), Box<dyn std::error::Error>> {
-    let Some(registry) = open_registry(base_dir) else {
-        println!("No registry found");
-        return Ok(());
-    };
-
-    let entries: Vec<&ShmemEntry> = registry.entries().iter().filter(|e| entry_visible(e)).collect();
-    let apps = app_names(registry);
-
-    let total = entries.len();
-    let alive = entries.iter().filter(|e| e.pids.any_alive()).count();
-    let dead = total - alive;
-    let poisoned = entries.iter().filter(|e| check_poison(e).is_some()).count();
-
-    let mut queues = 0usize;
-    let mut arrays = 0usize;
-    let mut data = 0usize;
-    let mut unknown = 0usize;
-    let mut total_capacity: u64 = 0;
-    let mut total_elem_bytes: u64 = 0;
-
-    for entry in &entries {
-        match entry.kind {
-            ShmemKind::Queue => queues += 1,
-            ShmemKind::SeqlockArray => arrays += 1,
-            ShmemKind::Data => data += 1,
-            ShmemKind::Unknown => unknown += 1,
-        }
-        total_capacity += entry.capacity as u64;
-        total_elem_bytes += (entry.elem_size as u64) * (entry.capacity as u64);
-    }
-
-    let color = std::io::stdout().is_terminal();
-
-    let title = "flux shmem stats";
-    if color {
-        println!("{}", title.bold().underlined());
-    } else {
-        println!("{title}");
-    }
-    println!();
-
-    println!("  Apps:             {}", apps.len());
-    println!("  Total segments:   {total}");
-
-    if color {
-        println!("  Alive:            {}", format!("{alive}").green());
-        println!("  Dead:             {}", format!("{dead}").red());
-        if poisoned > 0 {
-            println!("  Poisoned:         {}", format!("{poisoned}").yellow());
-        } else {
-            println!("  Poisoned:         {poisoned}");
-        }
-    } else {
-        println!("  Alive:            {alive}");
-        println!("  Dead:             {dead}");
-        println!("  Poisoned:         {poisoned}");
-    }
-
-    println!();
-    println!("  By kind:");
-    println!("    Queue:          {queues}");
-    println!("    SeqlockArray:   {arrays}");
-    println!("    Data:           {data}");
-    if unknown > 0 {
-        println!("    Unknown:        {unknown}");
-    }
-
-    println!();
-    println!("  Total slots:      {total_capacity}");
-    println!(
-        "  Estimated memory: {}",
-        format_bytes(total_elem_bytes)
-    );
-
-    Ok(())
-}
-
-pub fn format_bytes(bytes: u64) -> String {
-    const KIB: u64 = 1024;
-    const MIB: u64 = 1024 * KIB;
-    const GIB: u64 = 1024 * MIB;
-
-    if bytes >= GIB {
-        format!("{:.1} GiB", bytes as f64 / GIB as f64)
-    } else if bytes >= MIB {
-        format!("{:.1} MiB", bytes as f64 / MIB as f64)
-    } else if bytes >= KIB {
-        format!("{:.1} KiB", bytes as f64 / KIB as f64)
-    } else {
-        format!("{bytes} B")
-    }
-}
-
-pub fn inspect(
-    base_dir: &Path,
-    app_filter: Option<&str>,
-    segment_filter: Option<&str>,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let Some(registry) = open_registry(base_dir) else {
-        println!("No registry found");
-        return Ok(());
-    };
-
-    for entry in registry.entries() {
-        if !entry_visible(entry) {
-            continue;
-        }
-        if let Some(app) = app_filter
-            && entry.app_name.as_str() != app
-        {
-            continue;
-        }
-        if let Some(seg) = segment_filter
-            && !entry.type_name.as_str().contains(seg)
-        {
-            continue;
-        }
-
-        let alive = entry.pids.any_alive();
-        println!("─── {} ───", entry.type_name.as_str());
-        println!("  App:       {}", entry.app_name.as_str());
-        println!("  Kind:      {}", entry.kind);
-        let active = entry.pids.active_pids();
-        if active.len() == 1 {
-            println!(
-                "  PID:       {} {}",
-                active[0],
-                if alive { "(alive)" } else { "(dead)" }
-            );
-        } else {
-            println!("  PIDs:      {} attached", active.len());
-            for &pid in &active {
-                let s = if is_pid_alive(pid) { "alive" } else { "dead" };
-                println!("             {} ({})", pid, s);
-            }
-        }
-        println!("  Elem size: {} bytes", entry.elem_size);
-        println!("  Capacity:  {}", entry.capacity);
-        println!("  Flink:     {}", entry.flink.as_str());
-        println!("  Type hash: 0x{:016x}", entry.type_hash);
-
-        if entry.kind == ShmemKind::Queue
-            && let Ok(header) = QueueHeader::open_shared(entry.flink.as_str())
-        {
-            println!("  Queue type:   {:?}", header.queue_type);
-            println!(
-                "  Total writes: {}",
-                header.count.load(Ordering::Relaxed)
-            );
-            println!(
-                "  Mask:         {} (capacity={})",
-                header.mask,
-                header.mask + 1
-            );
-        }
-        println!();
-    }
-    Ok(())
-}
-
-pub fn clean(
-    base_dir: &Path,
-    app_filter: Option<&str>,
-    force: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let Some(registry) = open_registry(base_dir) else {
-        println!("No registry found");
-        return Ok(());
-    };
-
-    let mut stale = Vec::new();
-    for entry in registry.entries() {
-        if entry.is_empty() {
-            continue;
-        }
-        if let Some(app) = app_filter
-            && entry.app_name.as_str() != app
-        {
-            continue;
-        }
-        if !entry.pids.any_alive() && flink_reachable(entry.flink.as_str()) {
-            stale.push(entry);
-        }
-    }
-
-    if stale.is_empty() {
-        println!("No stale segments found");
-        return Ok(());
-    }
-
-    println!("Found {} stale segments:", stale.len());
-    for entry in &stale {
-        let pids = format_pids(entry);
-        println!(
-            "  💀 {} {} ({})",
-            entry.app_name.as_str(),
-            entry.type_name.as_str(),
-            pids,
-        );
-        if force {
-            let flink_path = Path::new(entry.flink.as_str());
-            match cleanup_flink(flink_path) {
-                Ok(()) => println!("    ✓ cleaned {}", flink_path.display()),
-                Err(e) => eprintln!("    ✗ {e}"),
-            }
-        }
-    }
-
-    if force {
-        let compacted = registry.compact();
-        if compacted > 0 {
-            println!("  Compacted registry: removed {compacted} empty slots");
-        }
-    } else {
-        println!("\nDry run. Use --force to actually remove flink files.");
-    }
-    Ok(())
-}
-
-// ─── Poison detection ───────────────────────────────────────────────────────
-
-#[derive(Clone, Debug)]
-pub struct PoisonInfo {
-    pub n_poisoned: usize,
-    pub first_slot: usize,
-    pub total_slots: usize,
-}
-
-/// Scan a segment's seqlock buffer for poisoned slots.
-///
-/// Works for both `Queue` and `SeqlockArray` — both use the same pattern:
-/// a `repr(C, align(64))` header followed by `Seqlock<T>` slots where the
-/// first 8 bytes of each slot are the `AtomicU64` version.
-///
-/// A slot is poisoned when its version is odd (write-in-progress) and stays
-/// odd for >10µs, meaning the writer crashed mid-write.
-pub fn check_poison(entry: &ShmemEntry) -> Option<PoisonInfo> {
-    match entry.kind {
-        ShmemKind::Queue => check_queue_poison(entry.flink.as_str()),
-        ShmemKind::SeqlockArray => check_array_poison(entry.flink.as_str()),
-        _ => None,
-    }
-}
-
-fn check_queue_poison(flink: &str) -> Option<PoisonInfo> {
-    const HEADER_SIZE: usize = 64; // QueueHeader is repr(C, align(64))
-
-    let shmem = ShmemConf::new().flink(flink).open().ok()?;
-    let base = shmem.as_ptr();
-    let shmem_len = shmem.len();
-
-    if shmem_len < std::mem::size_of::<QueueHeader>() {
-        std::mem::forget(shmem);
-        return None;
-    }
-
-    let header = unsafe { &*(base as *const QueueHeader) };
-    if !header.is_initialized() || header.elsize == 0 {
-        std::mem::forget(shmem);
-        return None;
-    }
-
-    let n_slots = header.mask + 1;
-    let elsize = header.elsize;
-
-    if HEADER_SIZE + n_slots * elsize > shmem_len {
-        std::mem::forget(shmem);
-        return None;
-    }
-
-    let buf_base = unsafe { base.add(HEADER_SIZE) };
-    let result = scan_seqlock_slots(buf_base, n_slots, elsize);
-    std::mem::forget(shmem);
-    result
-}
-
-fn check_array_poison(flink: &str) -> Option<PoisonInfo> {
-    const HEADER_SIZE: usize = 64; // ArrayHeader is repr(C, align(64))
-
-    let shmem = ShmemConf::new().flink(flink).open().ok()?;
-    let base = shmem.as_ptr();
-    let shmem_len = shmem.len();
-
-    if shmem_len < std::mem::size_of::<ArrayHeader>() {
-        std::mem::forget(shmem);
-        return None;
-    }
-
-    let header = unsafe { &*(base as *const ArrayHeader) };
-    if !header.is_initialized() || header.elsize == 0 {
-        std::mem::forget(shmem);
-        return None;
-    }
-
-    let n_slots = header.bufsize;
-    let elsize = header.elsize;
-
-    if HEADER_SIZE + n_slots * elsize > shmem_len {
-        std::mem::forget(shmem);
-        return None;
-    }
-
-    let buf_base = unsafe { base.add(HEADER_SIZE) };
-    let result = scan_seqlock_slots(buf_base, n_slots, elsize);
-    std::mem::forget(shmem);
-    result
-}
-
-/// Walk `n_slots` seqlock slots starting at `buf_base` with stride `elsize`,
-/// reading the `AtomicU64` version at offset 0 of each slot.
-///
-/// Two snapshot reads separated by a yield — if a slot's version is odd in
-/// both reads it's stuck (poisoned). Transient odd values from in-flight
-/// writes resolve between snapshots. This avoids blocking the TUI thread.
-fn scan_seqlock_slots(buf_base: *mut u8, n_slots: usize, elsize: usize) -> Option<PoisonInfo> {
-    // First pass: collect slots with odd versions.
-    let mut odd_slots: Vec<usize> = Vec::new();
-    for i in 0..n_slots {
-        let version_ptr = unsafe { buf_base.add(i * elsize) } as *const AtomicU64;
-        let v = unsafe { &*version_ptr }.load(Ordering::Acquire);
-        if v & 1 != 0 {
-            odd_slots.push(i);
-        }
-    }
-
-    if odd_slots.is_empty() {
-        return None;
-    }
-
-    // Yield to let any in-flight writes complete.
-    std::thread::yield_now();
-
-    // Second pass: re-check only the odd slots.
-    let mut first_slot = None;
-    let mut n_poisoned = 0;
-    for &i in &odd_slots {
-        let version_ptr = unsafe { buf_base.add(i * elsize) } as *const AtomicU64;
-        let v = unsafe { &*version_ptr }.load(Ordering::Acquire);
-        if v & 1 != 0 {
-            n_poisoned += 1;
-            if first_slot.is_none() {
-                first_slot = Some(i);
-            }
-        }
-    }
-
-    if n_poisoned > 0 {
-        Some(PoisonInfo { n_poisoned, first_slot: first_slot.unwrap(), total_slots: n_slots })
-    } else {
-        None
-    }
-}
-
-// ─── Backing file size ──────────────────────────────────────────────────────
-
-/// Return the size (in bytes) of the shared memory backing file.
-///
-/// Opens the shmem read-only via its flink and returns `Shmem::len()`.
-/// Returns `None` if the shmem cannot be opened. The shmem is deliberately
-/// leaked (via `mem::forget`) to avoid unmapping memory that other processes
-/// may still be using.
-pub fn backing_file_size(flink: &str) -> Option<u64> {
-    let shmem = ShmemConf::new().flink(flink).open().ok()?;
-    let size = shmem.len() as u64;
-    std::mem::forget(shmem);
-    Some(size)
-}
-
-// ─── Helpers ────────────────────────────────────────────────────────────────
-
-pub fn flink_reachable(flink: &str) -> bool {
-    if flink.is_empty() {
-        return false;
-    }
-    // Check that the flink file exists and is non-empty.  The shared_memory
-    // crate writes the OS ID into the flink file, so a valid (reachable) flink
-    // is always non-empty.  Previously this called `ShmemConf::open()` which
-    // leaked the mmap on success (the returned `Shmem` was dropped without
-    // being the owner, so the mapping was never unmapped).
-    std::fs::metadata(flink)
-        .map(|m| m.len() > 0)
-        .unwrap_or(false)
-}
-
-/// An entry is visible if it has live processes or its backing shmem still exists.
-pub fn entry_visible(entry: &ShmemEntry) -> bool {
-    if entry.is_empty() {
-        return false;
-    }
-    if entry.pids.any_alive() {
-        return true;
-    }
-    flink_reachable(entry.flink.as_str())
-}
-
-fn format_pids(entry: &ShmemEntry) -> String {
-    let active = entry.pids.active_pids();
-    match active.len() {
-        0 => "pid=none".into(),
-        1 => format!("pid={}", active[0]),
-        n => {
-            let list: Vec<String> = active.iter().map(|p| p.to_string()).collect();
-            format!("pids({n})=[{}]", list.join(","))
-        }
-    }
-}
-
-/// Read AT_CLKTCK from /proc/self/auxv (ELF auxiliary vector).
-/// Falls back to 100 if the file can't be read.
-fn clock_ticks_per_sec() -> u64 {
-    const AT_CLKTCK: u64 = 17;
-    let Ok(data) = std::fs::read("/proc/self/auxv") else {
-        return 100;
-    };
-    // auxv is a sequence of (u64 type, u64 value) pairs, terminated by AT_NULL=0.
-    for chunk in data.chunks_exact(16) {
-        let a_type = u64::from_ne_bytes(chunk[..8].try_into().unwrap());
-        if a_type == 0 {
-            break;
-        }
-        if a_type == AT_CLKTCK {
-            return u64::from_ne_bytes(chunk[8..16].try_into().unwrap());
-        }
-    }
-    100
-}
-
-fn read_start_time(pid: u32) -> Option<String> {
-    let stat = std::fs::read_to_string(format!("/proc/{pid}/stat")).ok()?;
-    let after_comm = stat.rsplit_once(')')?.1;
-    let fields: Vec<&str> = after_comm.split_whitespace().collect();
-    // Field 19 after comm = starttime (field 21 in full stat, 0-indexed)
-    let starttime_ticks: u64 = fields.get(19)?.parse().ok()?;
-    let ticks_per_sec = clock_ticks_per_sec();
-
-    let proc_stat = std::fs::read_to_string("/proc/stat").ok()?;
-    let btime_line = proc_stat.lines().find(|l| l.starts_with("btime "))?;
-    let btime_secs: u64 = btime_line.split_whitespace().nth(1)?.parse().ok()?;
-
-    let start_secs = btime_secs + starttime_ticks / ticks_per_sec;
-    let ts = std::time::UNIX_EPOCH + std::time::Duration::from_secs(start_secs);
-    Some(humantime::format_rfc3339_seconds(ts).to_string())
-}
